+++ components/src/main/java/uk/gov/hmrc/components/organism/information/InformationMessageCardView.kt
@@ -90,7 +90,9 @@ class InformationMessageCardView @JvmOverloads constructor(
         binding.contentBody.visibility = if (body != null) View.VISIBLE else View.GONE
     }
 
-    fun addHeadlineButtons(buttons: List<SecondaryButton>) {
+    fun setHeadlineButtons(buttons: List<SecondaryButton>) {
+        binding.headlineButtonsContainer.removeAllViews()
+
         val spacing8 = context.resources.getDimensionPixelSize(R.dimen.hmrc_spacing_8)
         val spacing16 = context.resources.getDimensionPixelSize(R.dimen.hmrc_spacing_16)
 
@@ -104,7 +106,9 @@ class InformationMessageCardView @JvmOverloads constructor(
         binding.viewBottomPadding.visibility = View.GONE
     }
 
-    fun addContentButtons(buttons: List<Button>) {
+    fun setContentButtons(buttons: List<Button>) {
+        binding.buttonsContainer.removeAllViews()
+
         val spacing8 = context.resources.getDimensionPixelSize(R.dimen.hmrc_spacing_8)
         val spacing16 = context.resources.getDimensionPixelSize(R.dimen.hmrc_spacing_16)